Fix a probable typo in Apple's code?
Maybe they changed `DERParseSequenceContent` after making libDER closed-source, and therefore this would be an issue with us using the last released libDER sources.

This fixes certificate parsing and it was just a simple reversion to the old function call, which is why I think it was a typo in their updated code.